Libs v12.2.2
I noticed that the parent process does not take a thread slot.
I was limited to 87% CPU usage opposed to 99% CPU usage when I had 7 going to 8 extended threads. 
I thought I needed to subtract a thread because the parent process counted as a thread. 
That is not the case, you must spawn N threads if you have N threads (no subtracting one due to parent process).